# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SoberanIA Labs Local RAG",
    "description": "Local-first RAG API with optional Ollama/Gemma integration.",
    "version": "0.3.1"
  },
  "paths": {
...
    "/api/runtime/test": {
      "post": {
        "summary": "Test Runtime",
        "operationId": "test_runtime_api_runtime_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuntimeTestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuntimeTestResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/local-data": {
      "delete": {
        "summary": "Reset Local Data",
        "operationId": "reset_local_data_api_local_data_delete",
        "parameters": [
          {
            "name": "X-Confirm-Local-Data-Reset",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Confirm-Local-Data-Reset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LocalDataResetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents": {
      "get": {
        "summary": "List Documents",
        "operationId": "list_documents_api_documents_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Document",
        "operationId": "create_document_api_documents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/upload": {
      "post": {
        "summary": "Upload Document",
        "operationId": "upload_document_api_documents_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}": {
      "delete": {
        "summary": "Delete Document",
        "operationId": "delete_document_api_documents__document_id__delete",
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "summary": "Chat",
        "operationId": "chat_api_chat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Body_upload_document_api_documents_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_documents_upload_post"
      },
...
      "ChatRequest": {
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Question"
          },
          "conversation_context": {
            "items": {
              "$ref": "#/components/schemas/ConversationMessage"
            },
            "type": "array",
            "maxItems": 12,
            "title": "Conversation Context"
          },
          "collection_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Id"
          },
          "top_k": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Top K"
          },
          "runtime_options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeOptions"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "question"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceChunk"
            },
            "type": "array",
            "title": "Sources"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "retrieval_query": {
            "type": "string",
            "title": "Retrieval Query"
          },
          "retrieval_top_k": {
            "type": "integer",
            "title": "Retrieval Top K"
          },
          "retrieval_mode": {
            "type": "string",
            "enum": [
              "dense",
              "hybrid"
            ],
            "title": "Retrieval Mode"
          },
          "collection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Id"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "sources",
          "provider",
          "model",
          "retrieval_query",
          "retrieval_top_k",
          "retrieval_mode",
          "latency_ms"
        ],
        "title": "ChatResponse"
      },
...
      "ConversationMessage": {
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "role",
          "content"
        ],
        "title": "ConversationMessage"
      },
      "DocumentCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Title"
          },
          "content": {
            "type": "string",
            "maxLength": 1000000,
            "minLength": 10,
            "title": "Content"
          },
          "source_type": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Source Type",
            "default": "manual"
          }
        },
        "type": "object",
        "required": [
          "title",
          "content"
        ],
        "title": "DocumentCreate"
      },
      "DocumentListResponse": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentResponse"
            },
            "type": "array",
            "title": "Documents"
          }
        },
        "type": "object",
        "required": [
          "documents"
        ],
        "title": "DocumentListResponse"
      },
      "DocumentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type"
          },
          "total_chars": {
            "type": "integer",
            "title": "Total Chars"
          },
          "total_chunks": {
            "type": "integer",
            "title": "Total Chunks"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "source_type",
          "total_chars",
          "total_chunks",
          "created_at",
          "updated_at"
        ],
        "title": "DocumentResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LocalDataResetResponse": {
        "properties": {
          "documents_deleted": {
            "type": "integer",
            "title": "Documents Deleted"
          },
          "chunks_deleted": {
            "type": "integer",
            "title": "Chunks Deleted"
          },
          "chat_messages_deleted": {
            "type": "integer",
            "title": "Chat Messages Deleted"
          }
        },
        "type": "object",
        "required": [
          "documents_deleted",
          "chunks_deleted",
          "chat_messages_deleted"
        ],
        "title": "LocalDataResetResponse"
      },
...
      "RuntimeOptions": {
        "properties": {
          "profile": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "economy",
                  "balanced",
                  "strong",
                  "custom"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile"
          },
          "model": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "num_ctx": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 32768.0,
                "minimum": 512.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Ctx"
          },
          "num_gpu": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 256.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Gpu"
          },
          "keep_alive": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 24
              },
              {
                "type": "null"
              }
            ],
            "title": "Keep Alive"
          },
          "temperature": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 2.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Temperature"
          }
        },
        "type": "object",
        "title": "RuntimeOptions"
      },
      "RuntimeTestRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Prompt",
            "default": "Responda apenas: ok"
          },
          "runtime_options": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RuntimeOptions"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "RuntimeTestRequest"
      },
      "RuntimeTestResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "success",
          "provider",
          "model",
          "latency_ms"
        ],
        "title": "RuntimeTestResponse"
      },
      "SourceChunk": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "document_id": {
            "type": "string",
            "title": "Document Id"
          },
          "document_title": {
            "type": "string",
            "title": "Document Title"
          },
          "chunk_index": {
            "type": "integer",
            "title": "Chunk Index"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "collection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Id"
          },
          "page_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Number"
          },
          "section_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Title"
          },
          "slide_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slide Number"
          },
          "sheet_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sheet Name"
          },
          "cell_range": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cell Range"
          },
          "source_locator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Locator"
          },
          "dense_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dense Score"
          },
          "dense_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dense Rank"
          },
          "lexical_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lexical Rank"
          },
          "fusion_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fusion Score"
          },
          "retrieval_channels": {
            "items": {
              "type": "string",
              "enum": [
                "dense",
                "lexical"
              ]
            },
            "type": "array",
            "title": "Retrieval Channels"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "document_title",
          "chunk_index",
          "score",
          "content"
        ],
        "title": "SourceChunk"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}